requirements.txt
streamlit
google-generativeai
requests
Pillow
audio-recorder-streamlit
streamlit-geolocation